Add JeOS minimal guest
After exploring some possibilities, we came up with a
minimal Fedora install, which presents an interesting
compromise between features and disk space. The idea
is to host the compressed image on my FedoraPeople
area and have code to automatically download it if it's
missing, and subsequently cache it.

The unattended install variant is in fact only for
people that, for some weird reason want to rebuild
the minimal guest on a convenient way, nothing more
than that.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>